requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auth0-python==4.7.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==42.0.8
distro==1.9.0
Flask==2.2.5
Flask-Session==0.5.0
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Levenshtein==0.25.1
MarkupSafe==3.0.2
multidict==6.2.0
openai==1.35.0
packaging==24.2
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-Levenshtein==0.25.1
RapidFuzz==3.12.2
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.3
yarl==1.18.3
